create/refactor dropout test utility functions
